Refactor useWallet
